- in python 3.x, use int instead of long - long => int

+++ als_example.py
@@ -9,8 +9,7 @@ spark = SparkSession(sc)
 
 lines = spark.read.text("data/mllib/als/sample_movielens_ratings.txt").rdd
 parts = lines.map(lambda row: row.value.split("::"))
-ratingsRDD = parts.map(lambda p: Row(userId=int(p[0]), movieId=int(p[1]), rating=float(p[2]),
-                                     timestamp=long(p[3])))
+ratingsRDD = parts.map(lambda p: Row(userId=int(p[0]), movieId=int(p[1]), rating=float(p[2]), timestamp=int(p[3])))
 ratings = spark.createDataFrame(ratingsRDD)
 (training, test) = ratings.randomSplit([0.8, 0.2])
